Initial CAnimationPose commit
CPose seems to only be inherited by CAnimationPose and doesn't contain any members, so for now it's in the same file as CAnimationPose just for inheritence's sake--curiously, it's part of Mirage, not Animation.

Also, there's some very-early-alpha "havok" classes mapped out here which can be dynamically modified if you hook the right function. Rename or relocate these wherever you please.