Refactored bugprone "volatile" optimized functions
Before the migration of functions to .inl files (which was a good change, mind you), a few methods were marked as "volatile" if they contained assembly code in their function body. This was to prevent compiler optimizations from damaging optimized function calls by putting things in the incorrect addresses etc.

While this was previously done in some cases due to inexperience (I believe I wrote both the StateMachineBase methods and the PlayerSpeedContext ones--I absolutely wrote the latter at least), this was a viable strategy for nested classes as their type names could not be forward declared. With the .inl migration, this is no longer necessary.

That said, not ONLY is it not necessary, but the use of "inline" keywords combined with volatile caused serious instability for these optimized functions, crashing very often. For safety and consistency, all these methods have been changed to call functions normally, passing `this` as an argument on the stack. Testing with these changes produced no crashes from these methods under average circumstances.