Support for const Message& sending on the stack
Please genuinely consider adding this. As I mention in a comment, the current implementation for shared pointer messages allows for const references already, and that same functionality is now possible with stack-based messages without `const_cast`.

If this PR is denied on the grounds that breaking constness is off the table, then messages sent as shared pointers shouldn't be allowed to be const reference either. Just be aware there's a few mods that already pass shared pointer messages in-place that will break because of this change. If it was okay then, I don't see why it's not okay now.